fix(deployment): circumvent git-hook
I have a global pre-commot-hook which does not let me commit if
there is no newline on the end of the file, as many open source
projects complain (and they're right!) if I submit a PR and don't
follow this.

As the generated stuff from wintersmith often does not have
newlines, I do not want my git-hook to yell at me when deploying.